requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==1.10.7
markdown==3.5.1
langchain==0.0.351
langchain-community==0.0.10
langchain-openai==0.0.2
faiss-cpu==1.7.4
networkx==3.2.1
spacy==3.7.2
langsmith==0.1.1
scikit-learn==1.3.2
